+++ src/core/plugins/imports.ts
@@ -2,7 +2,14 @@ import * as alt from 'alt-server';
 import logger from '../server/utility/athenaLogger';
 import { SYSTEM_EVENTS } from '../shared/enums/system';
 
-const filePaths = ['./teleportEffect/teleporter', './heist/heist', './deathLog/deathLog'];
+const filePaths = [
+    // GAMEMODE IMPORTS
+    './teleportEffect/teleporter',
+    './heist/heist',
+    './deathLog/deathLog',
+    './actionMenu/actionMenu'
+    // YOUR IMPORTS -- BELOW HERE
+];
 
 export default async function loadImports() {
     logger.info(`Loading extras folder...`);